Update MainMenu.cs
Bug Fix (Audio System): Refactored the AudioManager to get its AudioSource components via code in the Awake function. This prevents a race condition during scene loading that was causing audio source references to be lost, which in turn fixes a bug where grid generation would fail when loading the GameScene from the MainMenu.